Change `oif` > `oifname` in the Examples
`oif` only works for already existing interfaces, if anyone
uses the example for a service that starts after the firewall
config is set, the nftables will fail. For example if you
setup a wireguard server, wireguard will be up after the nft
setup, and the `wg$` interface will not exist by the time nft
is starting.